+++ src/js/forms.js
@@ -1,0 +1,107 @@
+(function ($) {
+    EqUI.forms = {};
+
+    // Init
+    EqUI.forms.init = function() {
+        // Global vars
+        EqUI.forms.element = $('.btn');
+        EqUI.forms.input_selector = 'input[type=text], input[type=password], input[type=email], input[type=url], input[type=tel], input[type=number], input[type=search], textarea';
+
+        // Update Labels
+        $(EqUI.forms.input_selector).each(function(index, element) {
+            if ($(element).val().length > 0 || $(this).attr('placeholder') !== undefined || $(element)[0].validity.badInput === true) {
+                $(this).siblings('label, i').addClass('active');
+            }
+            else {
+                $(this).siblings('label, i').removeClass('active');
+            }
+        });
+
+        // Handle HTML5 autofocus
+        $('input[autofocus]').siblings('label, i').addClass('active');
+
+        // Add active if form auto complete
+        $(document).on('change', EqUI.forms.input_selector, function () {
+            if($(this).val().length !== 0 || $(this).attr('placeholder') !== undefined) {
+                $(this).siblings('label, i').addClass('active');
+            }
+            EqUI.forms.validate_field($(this));
+        });
+
+        // HTML DOM FORM RESET handling
+        $(document).on('reset', function(e) {
+            var formReset = $(e.target);
+            if (formReset.is('form')) {
+                formReset.find(EqUI.forms.input_selector).removeClass('valid').removeClass('invalid');
+                formReset.find(EqUI.forms.input_selector).each(function () {
+                    if ($(this).attr('value') === '') {
+                        $(this).siblings('label, i').removeClass('active');
+                    }
+                });
+
+                // Reset select
+                formReset.find('select.initialized').each(function () {
+                    var reset_text = formReset.find('option[selected]').text();
+                    formReset.siblings('input.select-dropdown').val(reset_text);
+                });
+            }
+        });
+
+        // Add active when element has focus
+        $(document).on('focus', EqUI.forms.input_selector, function () {
+            $(this).siblings('label, i').addClass('active');
+        });
+
+        $(document).on('blur', EqUI.forms.input_selector, function () {
+            var $inputElement = $(this);
+            if ($inputElement.val().length === 0 && $inputElement[0].validity.badInput !== true && $inputElement.attr('placeholder') === undefined) {
+                $inputElement.siblings('label, i').removeClass('active');
+            }
+            EqUI.forms.validate_field($inputElement);
+        });
+
+        // Textarea auto size
+        autosize($('textarea.eq-ui-textarea'));
+
+    };
+
+    // Update
+    EqUI.forms.update = function() {
+
+    };
+
+    // Validate
+    EqUI.forms.validate_field = function(object) {
+        var hasLength = object.attr('length') !== undefined;
+        var lenAttr = parseInt(object.attr('length'));
+        var len = object.val().length;
+
+        if (object.val().length === 0 && object[0].validity.badInput === false) {
+            if (object.hasClass('validate')) {
+                object.removeClass('valid');
+                object.removeClass('invalid');
+            }
+        }
+        else {
+            if (object.hasClass('validate')) {
+                // Check for character counter attributes
+                if ((object.is(':valid') && hasLength && (len < lenAttr)) || (object.is(':valid') && !hasLength)) {
+                    object.removeClass('invalid');
+                    object.addClass('valid');
+                }
+                else {
+                    object.removeClass('valid');
+                    object.addClass('invalid');
+                }
+            }
+        }
+    };
+
+    $(document).ready(function() {
+        // Init
+        EqUI.forms.init();
+
+        // Update
+        EqUI.forms.update();
+    });
+}( jQuery ));