+++ src/js/layout/side-nav.js
@@ -1,10 +1,11 @@
 (function ($) {
     EqUI.side_nav = {};
 
-    EqUI.side_nav.element = $('.eq-ui-side-nav');
-
     // Init
     EqUI.side_nav.init = function() {
+        // Global vars
+        EqUI.side_nav.element = $('.eq-ui-side-nav');
+
         $('.eq-ui-side-nav-toggle').on('click', function(){
             EqUI.side_nav.nav_toggle();
         });
@@ -15,37 +16,37 @@
         
         if (window.innerWidth > 768) {
 
-            if($('body').hasClass('eq-ui-side-nav-auto')){
-                $('body').removeClass("eq-ui-side-nav-hide");
-                $('body').addClass("eq-ui-side-nav-visible");
-            } else if(!$('body').hasClass('eq-ui-side-nav-visible') && !$('body').hasClass('eq-ui-side-nav-hide')) {
-                $('body').addClass("eq-ui-side-nav-visible");
+            if(EqUI.site.body.hasClass('eq-ui-side-nav-auto')){
+                EqUI.site.body.removeClass("eq-ui-side-nav-hide");
+                EqUI.site.body.addClass("eq-ui-side-nav-visible");
+            } else if(!EqUI.site.body.hasClass('eq-ui-side-nav-visible') && !EqUI.site.body.hasClass('eq-ui-side-nav-hide')) {
+                EqUI.site.body.addClass("eq-ui-side-nav-visible");
             }
         }
         else {
 
-            if($('body').hasClass('eq-ui-side-nav-auto')){
-                $('body').removeClass("eq-ui-side-nav-visible");
-                $('body').addClass("eq-ui-side-nav-hide");
-            } else if(!$('body').hasClass('eq-ui-side-nav-visible') && !$('body').hasClass('eq-ui-side-nav-hide')) {
-                $('body').addClass("eq-ui-side-nav-hide");
+            if(EqUI.site.body.hasClass('eq-ui-side-nav-auto')){
+                EqUI.site.body.removeClass("eq-ui-side-nav-visible");
+                EqUI.site.body.addClass("eq-ui-side-nav-hide");
+            } else if(!EqUI.site.body.hasClass('eq-ui-side-nav-visible') && !EqUI.site.body.hasClass('eq-ui-side-nav-hide')) {
+                EqUI.site.body.addClass("eq-ui-side-nav-hide");
             }
         }
     };
 
     // Nav Toggle
     EqUI.side_nav.nav_toggle = function() {
-        if($('body').hasClass('eq-ui-side-nav-visible')){
-            $('body').removeClass("eq-ui-side-nav-visible");
-            $('body').addClass("eq-ui-side-nav-hide");
-        } else if($('body').hasClass('eq-ui-side-nav-hide')) {
-            $('body').removeClass("eq-ui-side-nav-hide");
-            $('body').addClass("eq-ui-side-nav-visible");
+        if(EqUI.site.body.hasClass('eq-ui-side-nav-visible')){
+            EqUI.site.body.removeClass("eq-ui-side-nav-visible");
+            EqUI.site.body.addClass("eq-ui-side-nav-hide");
+        } else if(EqUI.site.body.hasClass('eq-ui-side-nav-hide')) {
+            EqUI.site.body.removeClass("eq-ui-side-nav-hide");
+            EqUI.site.body.addClass("eq-ui-side-nav-visible");
         } else {
             if (window.innerWidth > 768) {
-                $('body').addClass("eq-ui-side-nav-hide");
+                EqUI.site.body.addClass("eq-ui-side-nav-hide");
             } else{
-                $('body').addClass("eq-ui-side-nav-visible");
+                EqUI.site.body.addClass("eq-ui-side-nav-visible");
             }
         }
     };