+++ src/js/site.js
@@ -6,9 +6,17 @@
 
     // Init
     EqUI.site.init = function() {
+        // Global vars
+        EqUI.site.body = $('body');
+
         // Is touch
         EqUI.site.isTouch = 'ontouchstart' in window || 'msmaxtouchpoints' in window.navigator;
         if(EqUI.site.isTouch){ $('html').addClass("is-touch"); }
+
+        // Set checkbox to indeterminate
+        $('[type="checkbox"].indeterminate-checkbox').each(function(index, element) {
+            element.indeterminate = true;
+        });
         
     };
 
@@ -16,8 +24,8 @@
     EqUI.site.update = function() {
         
         // Is layout header is fixed
-        if($('body').hasClass('eq-ui-layout-header-fixed')){
-            $('body').css('margin-top', $('.eq-ui-layout-header').outerHeight(true)+'px');
+        if(EqUI.site.body.hasClass('eq-ui-layout-header-fixed')){
+            EqUI.site.body.css('margin-top', $('.eq-ui-layout-header').outerHeight(true)+'px');
         }
         
         if (window.innerWidth > 768) {